Remove CollectionSerializer
The problem with CollectionSerializer is it creates a branching flow of logic.

Typically speaking, given an object, use the class's name to find the appropriate serializer. CollectionSerializer breaks this logic.

The alternative of embedding the array-like object serialization into the `#serialize` method itself seems to work nicely and carries the caller's namespace forward effectively.

That being said, my gut is telling me that there is a better way to handle this. TBD.