Add conditional attribute rendering
While implementing this feature, I realised there is a bit of logic to
the conditional rendering. This made `Serializer#as_json` bloat in
complexity.

To address the added complexity, I extracted out the `Attribute` and
`Association` classes.

This provides an added benefit of the `Attribute` and `Association`
classes defining the `#included?` method, which checks depth vs
max_depth, along with user defined
conditions.